+++ resources/views/back/menu/_partials/_rowitem.blade.php
@@ -9,6 +9,9 @@
             {{ $item->label }}
         @else
             <a href="{{ route('chief.back.menuitem.edit', $item->id) }}">{{ $item->label }}</a>
+            @if($item->hidden_in_menu || $item->draft || $item->archived)
+                <span class="text-subtle"><em>[OFFLINE]</em></span>
+            @endif
             @if($item->collection_type)
                 &nbsp;<em class="text-subtle">pagina groep</em>
             @endif