+++ resources/views/back/menu/show.blade.php
@@ -38,7 +38,6 @@
             <div class="stack-s">
 
                 @foreach($menuItem->children as $child)
-
                     @include('chief::back.menu._partials._rowitem', ['level' => 1, 'item' => $child])
 
                     <div class="stack-xs">