+++ tests/Feature/Authorization/AdminLoginTest.php
@@ -15,7 +15,7 @@ class AdminLoginTest extends TestCase
 {
     use ChiefDatabaseTransactions;
 
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();