+++ tests/Feature/Common/ArchiveTraitTest.php
@@ -21,7 +21,7 @@ class ArchivableTest extends TestCase
      */
     private $dummy;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();