+++ tests/Feature/Common/PublishableTraitTest.php
@@ -33,7 +33,7 @@ class PublishableTest extends TestCase
      */
     private $dummy;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();