+++ tests/Feature/Management/StoreManagerTest.php
@@ -12,7 +12,7 @@ class StoreManagerTest extends TestCase
 {
     private $fake;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();