+++ tests/Feature/Menu/CreateMenuItemTest.php
@@ -8,7 +8,7 @@ use Thinktomorrow\Chief\Pages\Page;
 
 class CreateMenuItemTest extends TestCase
 {
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();