+++ tests/Feature/Menu/DeleteMenuItemTest.php
@@ -7,7 +7,7 @@ use Thinktomorrow\Chief\Menu\MenuItem;
 
 class DeleteMenuItemTest extends TestCase
 {
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();