+++ tests/Feature/Modules/MorphableModuleTest.php
@@ -10,7 +10,7 @@ use Thinktomorrow\Chief\Tests\Fakes\NewsletterModuleFake;
 
 class MorphableModuleTest extends TestCase
 {
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
     }