+++ tests/Feature/PageBuilder/PageBuildTest.php
@@ -18,7 +18,7 @@ class PageBuildTest extends TestCase
 
     private $page;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();