+++ tests/Feature/Pages/ArchivePageTest.php
@@ -9,7 +9,7 @@ class ArchivePageTest extends TestCase
 {
     private $page;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();