+++ tests/Feature/Pages/CustomFieldsTest.php
@@ -17,7 +17,7 @@ class CustomFieldsTest extends TestCase
 
     private $page;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();