+++ tests/Feature/Sets/PageSetTest.php
@@ -11,7 +11,7 @@ use Thinktomorrow\Chief\Tests\TestCase;
 
 class PageSetTest extends TestCase
 {
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();