+++ tests/Feature/Sets/RelatingPageSetToPageTest.php
@@ -10,7 +10,7 @@ class RelatingPageSetToPageTest extends TestCase
 {
     private $page;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();