+++ tests/Feature/Users/ChangePasswordTest.php
@@ -11,7 +11,7 @@ class ChangePasswordTest extends TestCase
 {
     private $user;
 
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();