+++ tests/Feature/Users/DisableUserTest.php
@@ -9,7 +9,7 @@ class DisableUserTest extends TestCase
 {
     private $user;
 
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();