+++ tests/Feature/Users/InviteUserTest.php
@@ -12,7 +12,7 @@ use Illuminate\Support\Facades\Notification;
 
 class InviteUserTest extends TestCase
 {
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();