+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All Notable changes to the `chief` application template will be documented in this file. Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## unreleased
+- Fix: menuitems of offline pages now show in admin.
 
 ## 0.2.5 - 2019-03-05
 - Add: add `addTab()` method to `Thinktomorrow\Chief\Fields\FieldArrangement` class.